+++ packages/deploy-tooling/src/base/validate.ts
@@ -13,6 +13,7 @@ import {
 } from '@sap-ux/project-input-validator';
 import { EOL } from 'os';
 import type { AbapDeployConfig } from '../types';
+import { isAppStudio } from '@sap-ux/btp-utils';
 
 export type ValidationInputs = {
     appName: string;
@@ -21,6 +22,7 @@ export type ValidationInputs = {
     transport: string;
     client: string;
     url: string;
+    destination: string;
 };
 
 export type ValidationOutput = {
@@ -76,11 +78,13 @@ export async function validateBeforeDeploy(
         package: config.app.package ?? '',
         transport: config.app.transport ?? '',
         client: config.target.client ?? '',
-        url: config.target.url ?? ''
+        url: config.target.url ?? '',
+        destination: config.target.destination ?? ''
     };
 
     // output is passed by reference and status updated during the internal pipeline below.
     await validateInputTextFormat(input, output, provider, logger);
+    convertInputsForAdtValidations(input, output);
     await validatePackageWithAdt(input, output, provider, logger);
     await validateTransportRequestWithAdt(input, output, provider, logger);
 
@@ -89,6 +93,7 @@ export async function validateBeforeDeploy(
 
 /**
  * Format a list of summary records that is ready to be printed on the console.
+ * The reduce function makes sure a EOL is added at the beginning of the output.
  *
  * @param summary A list of summary records
  * @returns Formatted summary string
@@ -109,13 +114,37 @@ export function formatSummary(summary: SummaryRecord[]): string {
                     statusSymbol = yellow('?');
                     break;
             }
-            return `${'    '}${statusSymbol} ${next.message}`;
+            return `${' '.repeat(4)}${statusSymbol} ${next.message}`;
         })
         .reduce((aggregated, current) => {
             return `${aggregated}${EOL}${current}`;
         }, '');
 
     return summaryStr;
+}
+
+/**
+ *
+ * @param input
+ * @param output
+ */
+function convertInputsForAdtValidations(input: ValidationInputs, output: ValidationOutput): void {
+    const upperCasePackageName = input.package.toUpperCase();
+    const upperCaseTransport = input.transport.toUpperCase();
+    if (upperCasePackageName !== input.package) {
+        input.package = upperCasePackageName;
+        output.summary.push({
+            message: `Package name contains lower case letter(s). ${input.package} is used for ADT validation.`,
+            status: SummaryStatus.Unknown
+        });
+    }
+    if (upperCaseTransport !== input.transport) {
+        input.transport = upperCaseTransport;
+        output.summary.push({
+            message: `Transport request number contains lower case letter(s). ${input.transport} is used for ADT validation.`,
+            status: SummaryStatus.Unknown
+        });
+    }
 }
 
 /**
@@ -149,7 +178,7 @@ async function validateInputTextFormat(
     processInputValidationResult(result, output);
     result = validateClient(input.client);
     processInputValidationResult(result, output);
-    result = validateUrl(input.url);
+    result = validateUrlForOnPremTargetOnly(input.destination, input.url);
     processInputValidationResult(result, output);
 
     // If all the text validation passed, only show the following success message.
@@ -158,6 +187,25 @@ async function validateInputTextFormat(
             message: summaryMessage.allClientCheckPass,
             status: SummaryStatus.Valid
         });
+    }
+}
+
+/**
+ * A wrapper of validateUrl(). It uses same logic in system-access module's createAbapServiceProvider()
+ * function to determine the deploy target. Only validate URL for on-prem ABAP deploy target.
+ *
+ * @param destination
+ * @param url
+ * @returns
+ */
+function validateUrlForOnPremTargetOnly(destination: string, url: string): boolean | string {
+    if (isAppStudio() && destination) {
+        // No validation required for destination name
+        return true;
+    } else if (url) {
+        return validateUrl(url);
+    } else {
+        return 'Invalid deploy target';
     }
 }
 
@@ -238,10 +286,7 @@ async function validatePackageWithAdt(
     }
 
     // ADT expects input package
-    let inputPackage = input.package;
-    if (inputPackage.toUpperCase() === '$TMP') {
-        inputPackage = '$TMP';
-    }
+    const inputPackage = input.package;
 
     try {
         const adtService = await provider.getAdtService<ListPackageService>(ListPackageService);