+++ packages/deploy-tooling/test/unit/base/validate.test.ts
@@ -7,6 +7,11 @@ import { t } from '@sap-ux/project-input-validator/src/i18n';
 
 const nullLogger = new ToolsLogger({ transports: [new NullTransport()] });
 
+import { isAppStudio } from '@sap-ux/btp-utils';
+
+jest.mock('@sap-ux/btp-utils');
+const mockIsAppStudio = isAppStudio as jest.Mock;
+
 describe('deploy-test validation', () => {
     // default app for testing
     const app = {
@@ -71,9 +76,61 @@ describe('deploy-test validation', () => {
             );
             expect(output.result).toBe(false);
             const summaryStr = formatSummary(output.summary);
-            expect(summaryStr).toContain(`${red('×')} ${t('InvalidAppNameMultipleReason')}`);
-            expect(summaryStr).toContain(`${t('AbapInvalidAppNameLength', { length: name.length })}`);
-            expect(summaryStr).toContain(`${t('AbapInvalidAppName', { prefix })}`);
+            expect(summaryStr).toContain(`${red('×')} ${t('deploy.invalidAppNameMultipleReason')}`);
+            expect(summaryStr).toContain(`${t('deploy.abapInvalidAppNameLength', { length: name.length })}`);
+            expect(summaryStr).toContain(`${t('deploy.abapInvalidAppName', { prefix })}`);
+        });
+
+        test('Detect invalid deploy target', async () => {
+            const name = 'ZAPP1';
+            const prefix = 'Z';
+            mockedAdtService.listPackages.mockResolvedValueOnce(['TESTPACKAGE', 'MYPACKAGE']);
+            mockedAdtService.getTransportRequests.mockResolvedValueOnce([
+                { transportNumber: 'T000001' },
+                { transportNumber: 'T000002' },
+                { transportNumber: 'T000003' }
+            ]);
+            mockedAdtService.getAtoInfo.mockResolvedValueOnce({
+                developmentPrefix: prefix
+            });
+
+            const output = await validateBeforeDeploy(
+                {
+                    app: { ...app, name },
+                    target: { ...target, url: '' }
+                },
+                mockedProvider as any,
+                nullLogger
+            );
+            expect(output.result).toBe(false);
+            const summaryStr = formatSummary(output.summary);
+            expect(summaryStr).toContain(`${t('Invalid deploy target')}`);
+        });
+
+        test('Skip validating url if destination is provided', async () => {
+            mockIsAppStudio.mockReturnValueOnce(true);
+            const name = 'ZAPP1';
+            const prefix = 'Z';
+            const destination = 'TestDestination';
+            mockedAdtService.listPackages.mockResolvedValueOnce(['TESTPACKAGE', 'MYPACKAGE']);
+            mockedAdtService.getTransportRequests.mockResolvedValueOnce([
+                { transportNumber: 'T000001' },
+                { transportNumber: 'T000002' },
+                { transportNumber: 'T000003' }
+            ]);
+            mockedAdtService.getAtoInfo.mockResolvedValueOnce({
+                developmentPrefix: prefix
+            });
+
+            const output = await validateBeforeDeploy(
+                {
+                    app: { ...app, name },
+                    target: { ...target, destination, url: '' }
+                },
+                mockedProvider as any,
+                nullLogger
+            );
+            expect(output.result).toBe(true);
         });
 
         test('adtService error', async () => {
@@ -105,9 +162,9 @@ describe('deploy-test validation', () => {
             expect(output.result).toBe(false);
             const summaryStr = formatSummary(output.summary);
             expect(summaryStr).toContain(`${yellow('?')} ${summaryMessage.adtServiceUndefined} for AtoService`);
-            expect(summaryStr).toContain(`${red('×')} ${t('InvalidAppNameMultipleReason')}`);
-            expect(summaryStr).toContain(`${t('AbapInvalidAppNameLength', { length: name.length })}`);
-            expect(summaryStr).toContain(`${t('CharactersForbiddenInAppName')}`);
+            expect(summaryStr).toContain(`${red('×')} ${t('deploy.invalidAppNameMultipleReason')}`);
+            expect(summaryStr).toContain(`${t('deploy.abapInvalidAppNameLength', { length: name.length })}`);
+            expect(summaryStr).toContain(`${t('deploy.charactersForbiddenInAppName')}`);
         });
 
         test('getAtoInfo throws error', async () => {
@@ -135,9 +192,9 @@ describe('deploy-test validation', () => {
             expect(output.result).toBe(false);
             const summaryStr = formatSummary(output.summary);
             expect(summaryStr).toContain(`${yellow('?')} ${summaryMessage.atoAdtAccessError}`);
-            expect(summaryStr).toContain(`${red('×')} ${t('InvalidAppNameMultipleReason')}`);
-            expect(summaryStr).toContain(`${t('AbapInvalidAppNameLength', { length: name.length })}`);
-            expect(summaryStr).toContain(`${t('CharactersForbiddenInAppName')}`);
+            expect(summaryStr).toContain(`${red('×')} ${t('deploy.invalidAppNameMultipleReason')}`);
+            expect(summaryStr).toContain(`${t('deploy.abapInvalidAppNameLength', { length: name.length })}`);
+            expect(summaryStr).toContain(`${t('deploy.charactersForbiddenInAppName')}`);
         });
     });
 
@@ -181,6 +238,9 @@ describe('deploy-test validation', () => {
             expect(output.result).toBe(true);
             const summaryStr = formatSummary(output.summary);
             expect(summaryStr).toContain(`${green('√')} ${summaryMessage.packageCheckPass}`);
+            expect(summaryStr).toContain(
+                `${yellow('?')} Package name contains lower case letter(s). $TMP is used for ADT validation.`
+            );
         });
 
         test('Invalid package name', async () => {
@@ -297,6 +357,71 @@ describe('deploy-test validation', () => {
             expect(output.result).toBe(true);
             const summaryStr = formatSummary(output.summary);
             expect(summaryStr).toContain(`${green('√')} ${summaryMessage.transportNotRequired}`);
+        });
+
+        test('Valid package name - small case local package $tmp', async () => {
+            mockedAdtService.listPackages.mockResolvedValueOnce(['$TMP']);
+            mockedAdtService.getTransportRequests.mockRejectedValueOnce(
+                new Error(TransportChecksService.LocalPackageError)
+            );
+            mockedAdtService.getAtoInfo.mockResolvedValueOnce({
+                developmentPrefix: 'Z'
+            });
+
+            const output = await validateBeforeDeploy(
+                {
+                    app: { ...app, package: '$tmp' },
+                    target
+                },
+                mockedProvider as any,
+                nullLogger
+            );
+            expect(output.result).toBe(true);
+            expect(mockedAdtService.listPackages).toBeCalledWith({ 'phrase': '$TMP' });
+            expect(mockedAdtService.getTransportRequests).toBeCalledWith('$TMP', 'ZAPP1');
+            const summaryStr = formatSummary(output.summary);
+            expect(summaryStr).toContain(`${green('√')} ${summaryMessage.transportNotRequired}`);
+            expect(summaryStr).toContain(
+                `${yellow('?')} Package name contains lower case letter(s). $TMP is used for ADT validation.`
+            );
+        });
+
+        test('Valid transport - small case transport number', async () => {
+            mockedAdtService.listPackages.mockResolvedValueOnce(['TEST']);
+            mockedAdtService.getTransportRequests.mockResolvedValueOnce([
+                { transportNumber: 'T000001' },
+                { transportNumber: 'T000002' }
+            ]);
+            mockedAdtService.getAtoInfo.mockResolvedValueOnce({
+                developmentPrefix: 'Z'
+            });
+
+            const output = await validateBeforeDeploy(
+                {
+                    app: {
+                        ...testConfig.app,
+                        package: 'test',
+                        transport: 't000002'
+                    },
+                    target: { ...testConfig.target }
+                },
+                mockedProvider as any,
+                nullLogger
+            );
+            expect(output.result).toBe(true);
+            expect(mockedAdtService.listPackages).toBeCalledWith({ 'phrase': 'TEST' });
+            expect(mockedAdtService.getTransportRequests).toBeCalledWith('TEST', 'ZAPP1');
+            const summaryStr = formatSummary(output.summary);
+            expect(summaryStr).toContain(
+                `${yellow('?')} Package name contains lower case letter(s). TEST is used for ADT validation.`
+            );
+            expect(summaryStr).toContain(
+                `${yellow(
+                    '?'
+                )} Transport request number contains lower case letter(s). T000002 is used for ADT validation.`
+            );
+            expect(summaryStr).toContain(`${green('√')} ${summaryMessage.packageCheckPass}`);
+            expect(summaryStr).toContain(`${green('√')} ${summaryMessage.transportCheckPass}`);
         });
 
         test('adtService error', async () => {