+++ packages/preview-middleware-client/test/unit/adp/controllers/AddFragment.controller.test.ts
@@ -1,5 +1,5 @@
 import type Dialog from 'sap/m/Dialog';
-import type Event from 'sap/ui/base/Event';
+import Event from 'sap/ui/base/Event';
 import type UI5Element from 'sap/ui/core/Element';
 import JSONModel from 'sap/ui/model/json/JSONModel';
 import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
@@ -10,6 +10,7 @@ import { fetchMock, sapCoreMock } from 'mock/window';
 import ControlUtils from '../../../../src/adp/control-utils';
 import AddFragment from '../../../../src/adp/controllers/AddFragment.controller';
 import rtaMock from 'mock/sap/ui/rta/RuntimeAuthoring';
+import { ValueState } from 'mock/sap/ui/core/library';
 
 describe('AddFragment', () => {
     beforeAll(() => {
@@ -66,10 +67,18 @@ describe('AddFragment', () => {
 
             const openSpy = jest.fn();
             addFragment.byId = jest.fn().mockReturnValue({
-                open: openSpy
+                open: openSpy,
+                close: jest.fn(),
+                setEscapeHandler: jest.fn()
             });
 
+            addFragment.getView = jest.fn().mockReturnValue({ destroy: jest.fn(), setModel: jest.fn() });
+
             await addFragment.onInit();
+
+            const escapeHandlerCb = (addFragment.dialog.setEscapeHandler as jest.Mock).mock.calls[0][0];
+
+            escapeHandlerCb({ resolve: jest.fn() });
 
             expect(openSpy).toHaveBeenCalledTimes(1);
         });
@@ -176,12 +185,11 @@ describe('AddFragment', () => {
                 {} as unknown as RuntimeAuthoring
             );
 
-            const valueStateSpy = jest.fn();
+            const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
             const event = {
                 getSource: jest.fn().mockReturnValue({
                     getValue: jest.fn().mockReturnValue('Delete'),
-                    setValueState: valueStateSpy,
-                    setValueStateText: jest.fn()
+                    setValueState: valueStateSpy
                 })
             };
 
@@ -193,7 +201,7 @@ describe('AddFragment', () => {
 
             addFragment.onFragmentNameInputChange(event as unknown as Event);
 
-            expect(valueStateSpy).toHaveBeenCalledWith('Error');
+            expect(valueStateSpy).toHaveBeenCalledWith(ValueState.Error);
         });
 
         test('sets error when the fragment name is empty', () => {
@@ -203,12 +211,11 @@ describe('AddFragment', () => {
                 {} as unknown as RuntimeAuthoring
             );
 
-            const valueStateSpy = jest.fn();
+            const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
             const event = {
                 getSource: jest.fn().mockReturnValue({
                     getValue: jest.fn().mockReturnValue(''),
-                    setValueState: valueStateSpy,
-                    setValueStateText: jest.fn()
+                    setValueState: valueStateSpy
                 })
             };
 
@@ -220,7 +227,7 @@ describe('AddFragment', () => {
 
             addFragment.onFragmentNameInputChange(event as unknown as Event);
 
-            expect(valueStateSpy).toHaveBeenCalledWith('None');
+            expect(valueStateSpy).toHaveBeenCalledWith(ValueState.None);
         });
 
         test('sets error when the fragment name is has special characters', () => {
@@ -230,12 +237,11 @@ describe('AddFragment', () => {
                 {} as unknown as RuntimeAuthoring
             );
 
-            const valueStateSpy = jest.fn();
+            const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
             const event = {
                 getSource: jest.fn().mockReturnValue({
                     getValue: jest.fn().mockReturnValue('Share 2$5!'),
-                    setValueState: valueStateSpy,
-                    setValueStateText: jest.fn()
+                    setValueState: valueStateSpy
                 })
             };
 
@@ -247,8 +253,34 @@ describe('AddFragment', () => {
 
             addFragment.onFragmentNameInputChange(event as unknown as Event);
 
-            expect(valueStateSpy).toHaveBeenCalledWith('Error');
+            expect(valueStateSpy).toHaveBeenCalledWith(ValueState.Error);
         });
+
+        test('sets error when the fragment name exceeds 64 characters', () => {
+            const addFragment = new AddFragment(
+                'adp.extension.controllers.AddFragment',
+                {} as unknown as UI5Element,
+                {} as unknown as RuntimeAuthoring
+            );
+
+            const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
+            const event = {
+                getSource: jest.fn().mockReturnValue({
+                    getValue: jest.fn().mockReturnValue('thisisverylongnamethisisverylongnamethisisverylongnamethisisveryl'),
+                    setValueState: valueStateSpy
+                })
+            };
+
+            addFragment.model = testModel;
+            
+            addFragment.dialog = {
+                getBeginButton: jest.fn().mockReturnValue({ setEnabled: jest.fn() })
+            } as unknown as Dialog;
+
+            addFragment.onFragmentNameInputChange(event as unknown as Event);
+
+            expect(valueStateSpy).toHaveBeenCalledWith(ValueState.Error);
+        })
 
         test('sets create button to true when the fragment name is valid', () => {
             const addFragment = new AddFragment(
@@ -257,12 +289,11 @@ describe('AddFragment', () => {
                 {} as unknown as RuntimeAuthoring
             );
 
-            const valueStateSpy = jest.fn();
+            const valueStateSpy = jest.fn().mockReturnValue({ setValueStateText: jest.fn() });
             const event = {
                 getSource: jest.fn().mockReturnValue({
                     getValue: jest.fn().mockReturnValue('Share'),
-                    setValueState: valueStateSpy,
-                    setValueStateText: jest.fn()
+                    setValueState: valueStateSpy
                 })
             };
 
@@ -274,7 +305,7 @@ describe('AddFragment', () => {
 
             addFragment.onFragmentNameInputChange(event as unknown as Event);
 
-            expect(valueStateSpy).toHaveBeenCalledWith('None');
+            expect(valueStateSpy).toHaveBeenCalledWith(ValueState.Success);
         });
     });