+++ packages/preview-middleware-client/types/sap.ui.rta.d.ts
@@ -1,10 +1,21 @@
 declare module 'sap/ui/rta/command/BaseCommand' {
     import type Element from 'sap/ui/core/Element';
     import type ManagedObject from 'sap/ui/base/ManagedObject';
+    import type Component from 'sap/ui/core/Component';
+
+    type Selector = {
+        id: string;
+        controlType: string;
+        appComponent: Component;
+    };
 
     interface BaseCommand extends ManagedObject {
         execute(): Promise<void>;
         getElement(): Element;
+        getName(): string;
+        getSelector(): Selector;
+        getChangeType(): string;
+        getCommands(): BaseCommand[];
     }
 
     export default BaseCommand;
@@ -33,6 +44,19 @@ declare module 'sap/ui/rta/command/FlexCommand' {
     export default FlexCommand;
 }
 
+declare module 'sap/ui/rta/plugin/AddXMLAtExtensionPoint' {
+    import type CommandFactory from 'sap/ui/rta/command/CommandFactory';
+
+    interface Arguments {
+        commandFactory: CommandFactory;
+        fragmentHandler: (overlay: UI5Element, extensionPointInfo: uknown) => Promise<void | object>;
+    }
+
+    export default class AddXMLAtExtensionPoint {
+        constructor(_: Arguments) {}
+    }
+}
+
 declare module 'sap/ui/rta/command/CommandFactory' {
     import type BaseCommand from 'sap/ui/rta/command/BaseCommand';
     import type ManagedObject from 'sap/ui/base/ManagedObject';
@@ -40,8 +64,14 @@ declare module 'sap/ui/rta/command/CommandFactory' {
     import type Element from 'sap/ui/core/Element';
     import type { FlexSettings } from 'sap/ui/rta/RuntimeAuthoring';
 
-    interface CommandFactory {
-        getCommandFor<T extends BaseCommand = BaseCommand>(
+    interface Arguments {
+        flexSettings?: FlexSettings;
+    }
+
+    export default class CommandFactory {
+        constructor(_: Arguments) {}
+
+        static async getCommandFor<T extends BaseCommand = BaseCommand>(
             control: Element | ManagedObject | string,
             commandType: string,
             settings: object,
@@ -49,15 +79,12 @@ declare module 'sap/ui/rta/command/CommandFactory' {
             flexSettings?: FlexSettings
         ): Promise<T>;
     }
-
-    const CommandFactory: CommandFactory;
-    export default CommandFactory;
 }
 
 declare module 'sap/ui/rta/command/OutlineService' {
     export interface OutlineViewNode {
         id: string;
-        type: 'aggregation' | 'element';
+        type: 'aggregation' | 'element' | 'extensionPoint';
         technicalName: string;
         editable: boolean;
         elements?: OutlineViewNode[];
@@ -82,6 +109,7 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
     import type ElementOverlay from 'sap/ui/dt/ElementOverlay';
     import type ContextMenu from 'sap/ui/dt/plugin/ContextMenu';
     import type { Layer } from 'sap/ui/fl';
+    import type { Scenario } from 'sap/ui/fl/Scenario';
 
     type Manifest = {
         [key: string]: unknown;
@@ -119,7 +147,7 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
         /**
          * Key representing the current scenario
          */
-        scenario?: Scenario;
+        scenario: Scenario;
         /**
          * Generator of the change. Will be saved in the change.
          * This value is ignored by UI5 version prior to 1.107
@@ -135,7 +163,8 @@ declare module 'sap/ui/rta/RuntimeAuthoring' {
         getFlexSettings: () => FlexSettings;
         getService: <T>(name: 'outline' | 'controllerExtension' | string) => Promise<T>;
         getSelection: () => ElementOverlay[];
-        getDefaultPlugins: () => { contextMenu: ContextMenu };
+        getDefaultPlugins: () => { [key: string]: uknown; contextMenu: ContextMenu };
+        setPlugins: (defaultPlugins: object) => void;
         getRootControlInstance: () => {
             getManifest(): Manifest;
         } & Component;