+++ packages/ui-components/test/unit/components/UIDialog.test.tsx
@@ -283,7 +283,7 @@ describe('<UIDialog />', () => {
                   },
                   "backgroundColor": "var(--vscode-editorWidget-background)",
                   "border": "1px solid var(--vscode-editorWidget-border)",
-                  "borderRadius": 0,
+                  "borderRadius": 4,
                   "boxShadow": "0px 4px 10px rgba(0, 0, 0, 0.3)",
                   "minHeight": 100,
                   "overflow": "hidden",
@@ -350,7 +350,7 @@ describe('<UIDialog />', () => {
                 Object {
                   "backgroundColor": "var(--vscode-editorWidget-background)",
                   "border": "1px solid var(--vscode-editorWidget-border)",
-                  "borderRadius": 0,
+                  "borderRadius": 4,
                   "boxShadow": "0px 4px 10px rgba(0, 0, 0, 0.3)",
                   "minHeight": 100,
                 }