+++ examples/odata-cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @sap-ux/odata-cli
 
+## 0.12.4
+
+### Patch Changes
+
+-   Updated dependencies [db918804]
+    -   @sap-ux/axios-extension@1.7.2
+
+## 0.12.3
+
+### Patch Changes
+
+-   Updated dependencies [fa4537b2]
+    -   @sap-ux/axios-extension@1.7.1
+
 ## 0.12.2
 
 ### Patch Changes