+++ packages/control-property-editor/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux/control-property-editor
 
+## 0.4.1
+
+### Patch Changes
+
+-   e2b264c2: Make Control Property Editor aware which application (scenario) its running in the iframe
+
+## 0.4.0
+
+### Minor Changes
+
+-   6d2d2255: support all kind of changes from command stack
+
+## 0.3.1
+
+### Patch Changes
+
+-   c800f248: chore(deps): update dependency postcss to v8.4.31 [security]
+
+## 0.3.0
+
+### Minor Changes
+
+-   318e040e: Enables creation of XML fragments for Extension Points from the outline tree (when right-clicking on extension point) or from the application (when clicking on control).
+
 ## 0.2.4
 
 ### Patch Changes