+++ packages/deploy-tooling/CHANGELOG.md
@@ -1,5 +1,45 @@
 # @sap-ux/deploy-tooling
 
+## 0.11.6
+
+### Patch Changes
+
+-   58d48e13: Capitalize $tmp before performing any validations. ADT queries rejects package name that is not in capital letter.
+    The previous fix only caplitalize for list package query. But ADT query for valid transport requests also require
+    caplized capital letters to detect $TMP is local package.
+-   Updated dependencies [db918804]
+    -   @sap-ux/axios-extension@1.7.2
+    -   @sap-ux/system-access@0.3.4
+
+## 0.11.5
+
+### Patch Changes
+
+-   d0e46a5c: Do not validate url if destination is present
+-   Updated dependencies [d0e46a5c]
+    -   @sap-ux/project-input-validator@0.2.1
+
+## 0.11.4
+
+### Patch Changes
+
+-   70bf0627: When validation result is successful, the successful log messages should still be printed on the console.
+
+## 0.11.3
+
+### Patch Changes
+
+-   Updated dependencies [d31cfeff]
+    -   @sap-ux/project-input-validator@0.2.0
+
+## 0.11.2
+
+### Patch Changes
+
+-   Updated dependencies [fa4537b2]
+    -   @sap-ux/axios-extension@1.7.1
+    -   @sap-ux/system-access@0.3.3
+
 ## 0.11.1
 
 ### Patch Changes