+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @sap-ux-private/preview-middleware-client
 
+## 0.7.2
+
+### Patch Changes
+
+-   e2b264c2: Make Control Property Editor aware which application (scenario) its running in the iframe
+
+## 0.7.1
+
+### Patch Changes
+
+-   ca61803e: Fixed controller extension/fragment name longer than 64 chars error not showing up
+
+## 0.7.0
+
+### Minor Changes
+
+-   6d2d2255: support all kind of changes from command stack
+
+## 0.6.0
+
+### Minor Changes
+
+-   318e040e: Enables creation of XML fragments for Extension Points from the outline tree (when right-clicking on extension point) or from the application (when clicking on control).
+
 ## 0.5.1
 
 ### Patch Changes